Put singleton classes in their own file
This is to help a little with organization, and to avoid duplicating
code.
