+++ hexrd/ui/singletons.py
@@ -1,0 +1,25 @@
+from PySide2.QtCore import QObject
+
+
+class Singleton(type):
+
+    _instance = None
+
+    def __call__(cls, *args, **kwargs):
+        if cls._instance is None:
+            cls._instance = super().__call__(*args, **kwargs)
+
+        return cls._instance
+
+
+# This metaclass must inherit from `type(QObject)` for classes that use
+# it to inherit from QObject.
+class QSingleton(type(QObject)):
+
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+
+        return cls._instances[cls]
